requirements.txt
fastapi
uvicorn
streamlit
requests
python-multipart
pydantic
openai
autogen-agentchat
autogen-ext
sentence-transformers
psycopg2-binary
pgvector
pandas
numpy
pdfplumber
python-docx
scikit-learn
sqlalchemy==2.0.26
sqlalchemy-utils==0.42.1
tiktoken
autogen-ext[openai]
python-dotenv
asyncio